+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Add input related attributes for, max, min, maxlength, pattern placeholder and size
+
 ## [0.19.0] 2021-06-01
 
 - Add func Comment() for html comments